requirements/local.txt
-r base.txt
django-debug-toolbar==1.2.1
docutils==0.12
six==1.7.3
sqlparse==0.1.11
psycopg2==2.5.3
